0.4.0 - добавил интеграцию мета данных исходя из имени файла

+++ m_inside_meta/imside_meta.py
@@ -26,36 +26,37 @@ def insideMetaVID(basic_data):
 
     date = get_data_from_name(basic_data)
 
-    # Подготавливаем дату
-    date_object = datetime.strptime(date, "%Y-%m-%d_%H-%M-%S")
-    new_datetime = date_object.strftime("%Y-%m-%dT%H:%M:%S")
+    if date:
+        # Подготавливаем дату
+        date_object = datetime.strptime(date, "%Y-%m-%d_%H-%M-%S")
+        new_datetime = date_object.strftime("%Y-%m-%dT%H:%M:%S")
 
-    # Создаем временное имя файла для нового видео
-    temp_file = os.path.join(os.path.dirname(location_file), f"{name}_temp.{extension}")
+        # Создаем временное имя файла для нового видео
+        temp_file = os.path.join(os.path.dirname(location_file), f"{name}_temp.{extension}")
 
-    # Копируем видео с новыми метаданными
-    stream = ffmpeg.input(location_file)
-    stream = ffmpeg.output(
-        stream,
-        temp_file,  # Сохраняем во временный файл
-        vcodec='libx264',
-        preset='slow',
-        crf=18,
-        acodec='aac',
-        audio_bitrate='192k',
-        metadata=f'creation_time={new_datetime}',
-        c='copy',
-        y='-y'
-    )
+        # Копируем видео с новыми метаданными
+        stream = ffmpeg.input(location_file)
+        stream = ffmpeg.output(
+            stream,
+            temp_file,  # Сохраняем во временный файл
+            vcodec='libx264',
+            preset='slow',
+            crf=18,
+            acodec='aac',
+            audio_bitrate='192k',
+            metadata=f'creation_time={new_datetime}',
+            c='copy',
+            y='-y'
+        )
 
-    # Запускаем ffmpeg
-    ffmpeg.run(stream, overwrite_output=True)
+        # Запускаем ffmpeg
+        ffmpeg.run(stream, overwrite_output=True)
 
-    # Удаляем старое видео
-    os.remove(location_file)
+        # Удаляем старое видео
+        os.remove(location_file)
 
-    # Переименовываем временный файл в оригинальное имя
-    os.rename(temp_file, location_file)
+        # Переименовываем временный файл в оригинальное имя
+        os.rename(temp_file, location_file)
 
 
 def insideMetaIMG(basic_data):
@@ -68,39 +69,42 @@ def insideMetaIMG(basic_data):
     type_file = basic_data['type_file']
     location_file = basic_data['location_file']
 
+
     date = get_data_from_name(basic_data)
-    #print(date)
-    # Удаляем все мета данные из файла
-    # Открываем изображение
-    image = Image.open(location_file)
-    image.getexif().clear()
 
-    # Сохраняем изображение без метаданных во временный файл
-    temp_file = os.path.join(os.path.dirname(location_file), f"{name}_temp.{extension}")
-    image.save(temp_file)
+    if date:
+        #print(date)
+        # Удаляем все мета данные из файла
+        # Открываем изображение
+        image = Image.open(location_file)
+        image.getexif().clear()
 
-    # Открываем заново сохраненное изображение
-    image = Image.open(temp_file)
-    exif_dict = {"0th": {}, "Exif": {}, "GPS": {}, "Interop": {}, "1st": {}, "thumbnail": None}
+        # Сохраняем изображение без метаданных во временный файл
+        temp_file = os.path.join(os.path.dirname(location_file), f"{name}_temp.{extension}")
+        image.save(temp_file)
 
-    # Подготавливаем дату
-    date_object = datetime.strptime(date, "%Y-%m-%d_%H-%M-%S")
-    new_datetime = date_object.strftime("%Y:%m:%d %H:%M:%S")
+        # Открываем заново сохраненное изображение
+        image = Image.open(temp_file)
+        exif_dict = {"0th": {}, "Exif": {}, "GPS": {}, "Interop": {}, "1st": {}, "thumbnail": None}
 
-    exif_dict["Exif"][piexif.ExifIFD.DateTimeOriginal] = new_datetime
-    exif_dict["Exif"][piexif.ExifIFD.DateTimeDigitized] = new_datetime
+        # Подготавливаем дату
+        date_object = datetime.strptime(date, "%Y-%m-%d_%H-%M-%S")
+        new_datetime = date_object.strftime("%Y:%m:%d %H:%M:%S")
 
-    # Преобразуем метаданные в бинарную строку
-    exif_bytes = piexif.dump(exif_dict)
+        exif_dict["Exif"][piexif.ExifIFD.DateTimeOriginal] = new_datetime
+        exif_dict["Exif"][piexif.ExifIFD.DateTimeDigitized] = new_datetime
 
-    # Сохраняем изображение с метаданными во временный файл
-    image.save(temp_file, "jpeg", exif=exif_bytes)
+        # Преобразуем метаданные в бинарную строку
+        exif_bytes = piexif.dump(exif_dict)
 
-    # Удаляем старый файл
-    os.remove(location_file)
+        # Сохраняем изображение с метаданными во временный файл
+        image.save(temp_file, "jpeg", exif=exif_bytes)
 
-    # Переименовываем временный файл в оригинальное имя
-    os.rename(temp_file, location_file)
+        # Удаляем старый файл
+        os.remove(location_file)
+
+        # Переименовываем временный файл в оригинальное имя
+        os.rename(temp_file, location_file)
 
 
 def insideMeta(source_dir) -> None:
